Refactor EventOrganizer to split database and GraphQL layer
