+++ repository/event_organizer.go
@@ -3,22 +3,13 @@ package repository
 import (
 	"context"
 	"github.com/dotkom/a3s/ent"
-	"github.com/dotkom/a3s/graph/model"
 )
 
 type EventOrganizerRepository struct {
 	Client *ent.Client
 }
 
-func transform(eo *ent.EventOrganizer) *model.EventOrganizer {
-	return &model.EventOrganizer{
-		ID:    eo.ID,
-		Name:  eo.Name,
-		Email: eo.Email,
-	}
-}
-
-func (r *EventOrganizerRepository) Create(name string, email string) (*model.EventOrganizer, error) {
+func (r *EventOrganizerRepository) Create(name string, email string) (*ent.EventOrganizer, error) {
 	ctx := context.Background()
 	eventOrganizer, err := r.Client.EventOrganizer.Create().
 		SetName(name).
@@ -27,7 +18,7 @@ func (r *EventOrganizerRepository) Create(name string, email string) (*model.Eve
 	if err != nil {
 		return nil, err
 	}
-	return transform(eventOrganizer), nil
+	return eventOrganizer, nil
 }
 
 func (r *EventOrganizerRepository) Count() (int, error) {
@@ -39,26 +30,22 @@ func (r *EventOrganizerRepository) Count() (int, error) {
 	return count, nil
 }
 
-func (r *EventOrganizerRepository) All() ([]*model.EventOrganizer, error) {
+func (r *EventOrganizerRepository) All() ([]*ent.EventOrganizer, error) {
 	ctx := context.Background()
 	eventOrganizers, err := r.Client.EventOrganizer.Query().All(ctx)
 	if err != nil {
 		return nil, err
 	}
-	result := make([]*model.EventOrganizer, len(eventOrganizers))
-	for i, eo := range eventOrganizers {
-		result[i] = transform(eo)
-	}
-	return result, nil
+	return eventOrganizers, nil
 }
 
-func (r *EventOrganizerRepository) Get(id int) (*model.EventOrganizer, error) {
+func (r *EventOrganizerRepository) Get(id int) (*ent.EventOrganizer, error) {
 	ctx := context.Background()
 	eventOrganizer, err := r.Client.EventOrganizer.Get(ctx, id)
 	if err != nil {
 		return nil, err
 	}
-	return transform(eventOrganizer), nil
+	return eventOrganizer, nil
 }
 
 func (r *EventOrganizerRepository) Delete(id int) error {
