--- resolvers/event.resolvers.go
@@ -1,1 +1,0 @@
-package resolvers
