+++ resolvers/event_organizer.mutations.go
@@ -6,5 +6,9 @@ import (
 )
 
 func CreateEventOrganizer(r *mutationResolver, ctx context.Context, data model.CreateEventOrganizerInput) (*model.EventOrganizer, error) {
-	return r.EventOrganizerRepository.Create(data.Name, data.Email)
+	eventOrganizer, err := r.EventOrganizerRepository.Create(data.Name, data.Email)
+	if err != nil {
+		return nil, err
+	}
+	return r.Query().EventOrganizer(ctx, eventOrganizer.ID)
 }
